backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
# gunicorn==20.1.0
djoser==2.2.0
drf-extra-fields==3.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
Pillow==9.0.0
django-filter==23.2
